init: migrate app to next.js by using Pages folder API
