+++ app/components/MainPage.tsx
@@ -1,0 +1,102 @@
+import ErrorBoundary from '@/app/components/ErrorBoundary';
+import FlyoutBox from '@/app/components/FlyoutBox';
+import Pagination from '@/app/components/Pagination';
+import PeopleList from '@/app/components/PeopleList';
+import Search from '@/app/components/Search';
+import { PAGINATION_PAGE } from '@/app/constants';
+import store from '@/app/stores';
+import { setPage } from '@/app/stores/peopleSlice';
+import { RootState, useFetchPeopleQuery } from '@/app/stores/reducers';
+import { ThemeContext, themes } from '@/app/theme/ThemeContext';
+import ThemeProvider from '@/app/theme/ThemeProvider';
+import React from 'react';
+import { Provider, useDispatch, useSelector } from 'react-redux';
+
+// const Main = () => {
+//   <React.StrictMode>
+//     <Provider store={store}>
+
+export default function MainPage() {
+  const dispatch = useDispatch();
+  const { searchText, page, personList } = useSelector(
+    (state: RootState) => state.people
+  );
+  const { data, isFetching } = useFetchPeopleQuery({ searchText, page });
+  // const location = useLocation();
+  // const navigate = useNavigate();
+  const pageCount = data?.count ? Math.ceil(data?.count / PAGINATION_PAGE) : 0;
+
+  // function updateURLParams(params: { searchText?: string; page?: number }) {
+  //   const newParams = { ...queryString.parse(location.search), ...params };
+  //   navigate({ search: queryString.stringify(newParams) });
+  // }
+
+  function handleNextPageClick() {
+    if (pageCount && page < pageCount) {
+      const newPage = page + 1;
+      updateURLParams({ page: newPage });
+      dispatch(setPage(newPage));
+    }
+  }
+
+  function handlePrevPageClick() {
+    if (page > 1) {
+      const newPage = page - 1;
+      //updateURLParams({ page: newPage });
+      dispatch(setPage(newPage));
+    }
+  }
+
+  return (
+    <ThemeProvider>
+      <ErrorBoundary>
+        <div className="app">
+          <header className="header">
+            <p className="theme_selector">
+              Change theme
+              <ThemeContext.Consumer>
+                {({ theme, setTheme }) => (
+                  <i
+                    className="fa-solid fa-sun light"
+                    data-testid="theme-switch"
+                    onClick={() => {
+                      if (theme === themes.light) setTheme(themes.dark);
+                      if (theme === themes.dark) setTheme(themes.light);
+                    }}
+                  />
+                )}
+              </ThemeContext.Consumer>
+            </p>
+            <h1>Star Wars People Finders</h1>
+            <Search />
+          </header>
+          <main className="main">
+            <div className="people-wrapper">
+              {isFetching ? (
+                <div className="preloader" data-testid="loader"></div>
+              ) : (
+                <PeopleList people={data?.results} />
+              )}
+              <div className="person-detail">{/* <Outlet /> */}</div>
+            </div>
+            {!isFetching && pageCount > 1 && (
+              <Pagination
+                onNextPageClick={handleNextPageClick}
+                onPrevPageClick={handlePrevPageClick}
+                disable={{
+                  left: page === 1,
+                  right: page === data?.count,
+                }}
+                nav={{ current: page, total: pageCount }}
+              />
+            )}
+
+            {!isFetching && personList.length !== 0 && <FlyoutBox />}
+          </main>
+        </div>
+      </ErrorBoundary>
+    </ThemeProvider>
+  );
+}
+
+// export default Main;
