+++ app/dashboard/404.tsx
@@ -1,0 +1,10 @@
+const ErrorPage = () => {
+  return (
+    <div>
+      <h1>Oops!</h1>
+      <p>Sorry, an unexpected error has occurred.</p>
+    </div>
+  );
+};
+
+export default ErrorPage;
