+++ app/dashboard/details/page.tsx
