+++ app/test/DetailsPage.test.tsx
