+++ app/theme/ThemeProvider.tsx
@@ -2,17 +2,22 @@ import { useEffect, useState } from 'react';
 import { ThemeContext, themes } from './ThemeContext';
 import { ThemeProviderProps } from '../interfaces';
 
-const getTheme = () => {
-  const theme = `${localStorage.getItem('theme')}`;
-  if (Object.values(themes).includes(theme)) return theme;
+const getInitialTheme = () => {
+  if (typeof window === 'undefined') {
+    return themes.light;
+  }
 
-  const userMedia = window.matchMedia('(prefers-color-scheme:light)');
-  if (userMedia.matches) return themes.light;
-  return themes.dark;
+  const storedTheme = localStorage.getItem('theme');
+  if (storedTheme && Object.values(themes).includes(storedTheme)) {
+    return storedTheme;
+  }
+
+  const userMedia = window.matchMedia('(prefers-color-scheme: light)');
+  return userMedia.matches ? themes.light : themes.dark;
 };
 
 const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
-  const [theme, setTheme] = useState(getTheme);
+  const [theme, setTheme] = useState(getInitialTheme);
 
   useEffect(() => {
     document.documentElement.dataset.theme = theme;
