--- index.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" type="image/png" href="/favicon.png" />
-    <script
-      defer
-      src="https://kit.fontawesome.com/b0c0f87c06.js"
-      crossorigin="anonymous"
-    ></script>
-    <title>Star Wars People Finders</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
