+++ package.json
@@ -1,54 +1,37 @@
 {
-  "name": "hooks-and-routing",
+  "name": "app-name",
+  "version": "0.1.0",
   "private": true,
-  "version": "0.0.0",
-  "type": "module",
   "scripts": {
-    "dev": "vite --open",
-    "build": "tsc -b && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
     "format": "prettier --write .",
     "prepare": "husky",
     "test": "vitest run",
     "coverage": "vitest --coverage"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^2.2.6",
-    "query-string": "^9.0.0",
-    "react": "^18.3.1",
+    "@reduxjs/toolkit": "^2.2.7",
+    "next": "14.2.5",
+    "react": "^18",
     "react-csv": "^2.2.2",
-    "react-dom": "^18.3.1",
+    "react-dom": "^18",
     "react-redux": "^9.1.2",
-    "react-router-dom": "^6.24.1"
+    "redux": "^5.0.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.4.6",
-    "@testing-library/react": "^16.0.0",
-    "@types/node": "^20.14.9",
-    "@types/react": "^18.3.3",
+    "@testing-library/jest-dom": "^6.4.8",
+    "@types/node": "^20",
+    "@types/react": "^18",
     "@types/react-csv": "^1.1.10",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.15.0",
-    "@typescript-eslint/parser": "^7.15.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "@vitest/coverage-v8": "^2.0.2",
-    "c8": "^10.1.2",
-    "eslint": "^8.57.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.34.3",
-    "eslint-plugin-react-compiler": "^0.0.0-experimental-0998c1e-20240625",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "happy-dom": "^14.12.3",
-    "husky": "^9.0.11",
-    "msw": "^2.3.4",
-    "prettier": "3.3.2",
-    "redux-mock-store": "^1.5.4",
-    "typescript": "^5.2.2",
-    "vite": "^5.3.1",
-    "vitest": "^2.0.2"
+    "@types/react-dom": "^18",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.5",
+    "husky": "^9.1.4",
+    "prettier": "^3.3.3",
+    "typescript": "^5",
+    "vitest": "^2.0.5"
   }
 }
