+++ pages/_document.tsx
@@ -1,0 +1,24 @@
+import Document, { Html, Head, Main, NextScript } from 'next/document';
+
+class MyDocument extends Document {
+  render() {
+    return (
+      <Html>
+        <Head>
+          <link rel="icon" type="image/png" href="/favicon.png" />
+          <script
+            defer
+            src="https://kit.fontawesome.com/b0c0f87c06.js"
+            crossOrigin="anonymous"
+          ></script>
+        </Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    );
+  }
+}
+
+export default MyDocument;
