+++ pages/index.tsx
@@ -1,0 +1,7 @@
+import MainPage from '@/app/components/MainPage';
+
+const Home = () => {
+  return <MainPage />;
+};
+
+export default Home;
