--- vite.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vitest/config';
-import react from '@vitejs/plugin-react';
-/// <reference types="vitest" />
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  test: {
-    globals: true,
-    environment: 'happy-dom',
-    setupFiles: ['src/setupTest.ts'],
-  },
-});
